Refact: content script 리팩토링
- ClickMode, DragMode 분리
- speech, canvas, overlay 모듈 분리
