+++ client/src/background/index.ts
@@ -1,15 +1,9 @@
 chrome.runtime.onInstalled.addListener(() => {
   chrome.storage.sync.set({
-    clickMode: false,
-    dragMode: false,
+    isClickMode: false,
+    isDragMode: false,
   });
 });
-
-// chrome.runtime.onMessage.addListener((req, sender, res) => {
-//   chrome.tabs.captureVisibleTab((dataUrl) => {
-//     res({ dataUrl });
-//   });
-// });
 
 chrome.runtime.onConnect.addListener((port) => {
   chrome.tabs.captureVisibleTab((dataUrl) => {
